WMS_AI_PythonApp/WMS_LABEL_GEN/requirements.txt
supabase
python-dotenv
python-multipart
requests
playwright
fastapi
uvicorn
pydantic
